+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixed
 
+- `XIT ACT`: Fix an off-by-one error in the Refuel action
 - `XIT FINPR`: Fix profitability calculation after PRO license expiration
 - `cxpo-order-book`: Fix form label text overflow